requirements.txt
pipecat-ai[daily,openai,elevenlabs,silero]
python-dotenv
aiohttp
loguru
